added odds calc method to simulations
